+++ zone/entity.cpp
@@ -3514,7 +3514,7 @@ void EntityList::ClearZoneFeignAggro(Mob *targ)
 	}
 }
 
-void EntityList::AggroZone(Mob *who, uint64 hate)
+void EntityList::AggroZone(Mob *who, uint32 hate)
 {
 	auto it = npc_list.begin();
 	while (it != npc_list.end()) {
