+++ zone/questmgr.cpp
@@ -2023,9 +2023,9 @@ void QuestManager::clear_zone_flag(int zone_id) {
 	initiator->ClearZoneFlag(zone_id);
 }
 
-void QuestManager::sethp(int64 hpperc) {
+void QuestManager::sethp(int hpperc) {
 	QuestManagerCurrentQuestVars();
-	int64 newhp = (owner->GetMaxHP() * (100 - hpperc)) / 100;
+	int newhp = (owner->GetMaxHP() * (100 - hpperc)) / 100;
 	owner->Damage(owner, newhp, SPELL_UNKNOWN, EQ::skills::SkillHandtoHand, false, 0, false);
 }
