+++ zone/spells.cpp
@@ -2576,7 +2576,7 @@ bool Mob::SpellFinished(uint16 spell_id, Mob *spell_target, CastingSlot slot, ui
 	// lets not consume end for custom items that have disc procs.
 	// One might also want to filter out USE_ITEM_SPELL_SLOT, but DISCIPLINE_SPELL_SLOT are both #defined to the same thing ...
 	if (spells[spell_id].endurance_cost && !isproc) {
-		auto end_cost = (int64)spells[spell_id].endurance_cost;
+		auto end_cost = spells[spell_id].endurance_cost;
 		if (mgb)
 			end_cost *= 2;
 		SetEndurance(GetEndurance() - EQ::ClampUpper(end_cost, GetEndurance()));
@@ -2627,7 +2627,7 @@ bool Mob::SpellFinished(uint16 spell_id, Mob *spell_target, CastingSlot slot, ui
 				recast -= GetAA(aaTouchoftheWicked) * 720;
 			}
 
-			int64 reduction = CastToClient()->GetFocusEffect(focusReduceRecastTime, spell_id);
+			int reduction = CastToClient()->GetFocusEffect(focusReduceRecastTime, spell_id);
 
 			if (reduction) {
 				recast -= reduction;
@@ -4769,11 +4769,11 @@ float Mob::ResistSpell(uint8 resist_type, uint16 spell_id, Mob *caster, bool use
 		resist_modifier += caster->GetSpecialAbilityParam(CASTING_RESIST_DIFF, 0);
 	}
 
-	int64 focus_resist = caster->GetFocusEffect(focusResistRate, spell_id);
+	int focus_resist = caster->GetFocusEffect(focusResistRate, spell_id);
 
 	resist_modifier -= 2 * focus_resist;
 
-	int64 focus_incoming_resist = GetFocusEffect(focusFcResistIncoming, spell_id, caster);
+	int focus_incoming_resist = GetFocusEffect(focusFcResistIncoming, spell_id, caster);
 
 	resist_modifier -= focus_incoming_resist;
 
@@ -5808,7 +5808,7 @@ bool Mob::FindType(uint16 type, bool bOffensive, uint16 threshold) {
 				// adjustments necessary for offensive npc casting behavior
 				if (bOffensive) {
 					if (spells[buffs[i].spellid].effect_id[j] == type) {
-						int64 value =
+						int16 value =
 								CalcSpellEffectValue_formula(spells[buffs[i].spellid].buff_duration_formula,
 											spells[buffs[i].spellid].base_value[j],
 											spells[buffs[i].spellid].max_value[j],
@@ -6899,7 +6899,7 @@ bool Mob::IsFromTriggeredSpell(CastingSlot slot, uint32 item_slot) {
 	return false;
 }
 
-void Mob::SetHP(int64 hp)
+void Mob::SetHP(int32 hp)
 {
 	if (hp >= max_hp) {
 		current_hp = max_hp;
